@@ -1,64 +1,59 @@
-import router from './router'
-import store from './store'
-import {Message} from 'element-ui'
+import router from "@/router";
+import {useUserStoreHook} from "@/stores";
 import NProgress from 'nprogress' // progress bar
 import 'nprogress/nprogress.css' // progress bar style
-import {getToken} from '@/utils/auth' // get token from cookie
-import getPageTitle from '@/utils/get-page-title'
+import getPageTitle from "@/get-page-title";
+import {getToken} from "@/utils/auth";
+import {storeToRefs} from 'pinia'
+import {ElMessage} from "element-plus";
 
-NProgress.configure({showSpinner: false}) // NProgress Configuration
 
-const whiteList = ['/login'] // no redirect whitelist
+// 白名单
+const whiteList = ['/login', '/register']
 
+/** 路由前置守卫 */
 router.beforeEach(async (to, from, next) => {
-  // start progress bar
-  NProgress.start()
+    // start progress bar
+    // NProgress.start()
 
-  // set page title
-  document.title = getPageTitle(to.meta.title)
+    document.title = getPageTitle(to.meta.title)
 
-  // determine whether the user has logged in
-  const hasToken = getToken()
-
-  if (hasToken) {
-    if (to.path === '/login') {
-      // if is logged in, redirect to the home page
-      next({path: '/'})
-      NProgress.done()
-    } else {
-      const hasGetUserInfo = store.getters.name
-      if (hasGetUserInfo) {
-        next()
-      } else {
-        try {
-          // get user info
-          await store.dispatch('user/getInfo')
-
-          next()
-        } catch (error) {
-          // remove token and go to login page to re-login
-          await store.dispatch('user/resetToken')
-          Message.error(error || 'Has Error')
-          next(`/login?redirect=${to.path}`)
-          NProgress.done()
+    const hasToken = getToken()
+    // console.log(hasToken)
+    if (hasToken) {
+        if (to.path === '/login') {
+            next({path: '/'})
+            // NProgress.done()
+        } else {
+            const hasGetUserInfo = storeToRefs(useUserStoreHook().user_name)
+            if (hasGetUserInfo) {
+                next()
+            } else {
+                try {
+                    await useUserStoreHook.saveUserInfo()
+                    next()
+                } catch (err) {
+                    await useUserStoreHook.resetToken
+                    ElMessage({
+                        type: 'error',
+                        message: err || 'Has Error'
+                    })
+                    next(`/login?redirect=${to.path}`)
+                    // NProgress.done()
+                }
+            }
         }
-      }
-    }
-  } else {
-    /* has no token*/
-
-    if (whiteList.indexOf(to.path) !== -1) {
-      // in the free login whitelist, go directly
-      next()
     } else {
-      // other pages that do not have permission to access are redirected to the login page.
-      next(`/login?redirect=${to.path}`)
-      NProgress.done()
+        if (whiteList.indexOf(to.path) !== -1) {
+            next()
+        } else {
+            next(`/login?redirect=${to.path}`)
+            // NProgress.done()
+        }
     }
-  }
 })
 
 router.afterEach(() => {
-  // finish progress bar
-  NProgress.done()
+    // finish progress bar
+    NProgress.done()
 })
