@@ -1,181 +1,91 @@
-import Vue from 'vue'
-import Router from 'vue-router'
+import {createRouter, createWebHashHistory} from 'vue-router'
+import Layout from "@/layout/index.vue"
 
-Vue.use(Router)
-
-/* Layout */
-import Layout from '@/layout'
-
-/**
- * Note: sub-menu only appear when route children.length >= 1
- * Detail see: https://panjiachen.github.io/vue-element-admin-site/guide/essentials/router-and-nav.html
- *
- * hidden: true                   if set true, item will not show in the sidebar(default is false)
- * alwaysShow: true               if set true, will always show the root menu
- *                                if not set alwaysShow, when item has more than one children route,
- *                                it will becomes nested mode, otherwise not show the root menu
- * redirect: noRedirect           if set noRedirect will no redirect in the breadcrumb
- * name:'router-name'             the name is used by <keep-alive> (must set!!!)
- * meta : {
- roles: ['admin','editor']    control the page roles (you can set multiple roles)
- title: 'title'               the name show in sidebar and breadcrumb (recommend set)
- icon: 'svg-name'/'el-icon-x' the icon show in the sidebar
- breadcrumb: false            if set false, the item will hidden in breadcrumb(default is true)
- activeMenu: '/example/list'  if set path, the sidebar will highlight the path you set
- }
- */
-
-/**
- * constantRoutes
- * a base page that does not have permission requirements
- * all roles can be accessed
- */
 export const constantRoutes = [
-  {
-    path: '/login',
-      component: () => import('@/views/login/index'),
-      hidden: true
-  },
-
-  {
-      path: '/404',
-      component: () => import('@/views/404'),
-      hidden: true
-  },
-
-  {
-    path: '/',
-    component: Layout,
-      redirect: '/dashboard',
-      children: [{
-          path: 'dashboard',
-          name: 'Dashboard',
-          component: () => import('@/views/dashboard/index'),
-          meta: {title: 'Dashboard', icon: 'dashboard'}
-      }]
-  },
-
     {
-        path: '/example',
-        component: Layout,
-        redirect: '/example/table',
-        name: 'Example',
-        meta: {title: 'Example', icon: 'el-icon-s-help'},
-    children: [
-      {
-          path: 'table',
-          name: 'Table',
-          component: () => import('@/views/table/index'),
-          meta: {title: 'Table', icon: 'table'}
-      },
-        {
-            path: 'tree',
-            name: 'Tree',
-            component: () => import('@/views/tree/index'),
-            meta: {title: 'Tree', icon: 'tree'}
-      }
-    ]
-  },
-
-  {
-      path: '/form',
-    component: Layout,
-    children: [
-      {
-          path: 'index',
-          name: 'Form',
-          component: () => import('@/views/form/index'),
-          meta: {title: 'Form', icon: 'form'}
-      }
-    ]
-  },
-
-    {
-        path: '/nested',
-        component: Layout,
-        redirect: '/nested/menu1',
-        name: 'Nested',
+        path: '/login',
+        name: 'Login',
+        component: () => import('@/views/login/index.vue'),
+        hidden: true,
         meta: {
-            title: 'Nested',
-            icon: 'nested'
-        },
+            title: '登录',
+        }
+    },
+    {
+        path: '/register',
+        name: 'Register',
+        component: () => import('@/views/register/index.vue'),
+        hidden: true,
+        meta: {
+            title: '注册'
+        }
+    },
+    {
+        path: '/',
+        component: Layout,
+        redirect: '/home',
         children: [
             {
-                path: 'menu1',
-                component: () => import('@/views/nested/menu1/index'), // Parent router-view
-                name: 'Menu1',
-                meta: {title: 'Menu1'},
-                children: [
-                    {
-                        path: 'menu1-1',
-                        component: () => import('@/views/nested/menu1/menu1-1'),
-                        name: 'Menu1-1',
-                        meta: {title: 'Menu1-1'}
-                    },
-                    {
-                        path: 'menu1-2',
-                        component: () => import('@/views/nested/menu1/menu1-2'),
-                        name: 'Menu1-2',
-                        meta: {title: 'Menu1-2'},
-                        children: [
-                            {
-                                path: 'menu1-2-1',
-                                component: () => import('@/views/nested/menu1/menu1-2/menu1-2-1'),
-                                name: 'Menu1-2-1',
-                                meta: {title: 'Menu1-2-1'}
-                            },
-                            {
-                                path: 'menu1-2-2',
-                                component: () => import('@/views/nested/menu1/menu1-2/menu1-2-2'),
-                                name: 'Menu1-2-2',
-                                meta: {title: 'Menu1-2-2'}
-                            }
-                        ]
-                    },
-                    {
-                        path: 'menu1-3',
-                        component: () => import('@/views/nested/menu1/menu1-3'),
-                        name: 'Menu1-3',
-                        meta: {title: 'Menu1-3'}
-                    }
-                ]
+                path: "home",
+                name: "Home",
+                component: () => import("@/views/home/index.vue"),
+                meta: {title: 'Home', icon: 'Home'}
+            }
+        ]
+    },
+    {
+        path: '/article',
+        component: Layout,
+        redirect: '/article/list',
+        name: 'Article',
+        meta: {title: 'Article', icon: '<el-icon><Document /></el-icon>'},
+        children: [
+            {
+                path: 'list',
+                name: 'List',
+                component: () => import('@/views/article/article-list/index.vue'),
+                meta: {title: 'List', icon: '<el-icon><List /></el-icon>'}
             },
             {
-                path: 'menu2',
-                component: () => import('@/views/nested/menu2/index'),
-                name: 'Menu2',
-                meta: {title: 'menu2'}
-            }
-        ]
-    },
-
-    {
-        path: 'external-link',
-        component: Layout,
-        children: [
+                path: 'addArticle',
+                name: 'addArticle',
+                component: () => import('@/views/article/add-article/index.vue'),
+                meta: {title: 'addArticle', icon: '<el-icon><DocumentAdd /></el-icon>'}
+            },
             {
-                path: 'https://panjiachen.github.io/vue-element-admin-site/#/',
-                meta: {title: 'External Link', icon: 'link'}
+                path: 'tag',
+                name: 'tagManagement',
+                component: () => import('@/views/article/tag/index.vue'),
+                meta: {title: 'tagManagement', icon: '<el-icon><CollectionTag /></el-icon>'}
+            },
+            {
+                path: 'category',
+                name: 'category',
+                component: () => import('@/views/article/category/index.vue'),
+                meta: {title: 'category', icon: '<el-icon><PriceTag /></el-icon>'}
             }
         ]
-    },
-
-    // 404 page must be placed at the end !!!
-    {path: '*', redirect: '/404', hidden: true}
+    }
 ]
 
-const createRouter = () => new Router({
-    // mode: 'history', // require service support
-    scrollBehavior: () => ({y: 0}),
-    routes: constantRoutes
-})
-
-const router = createRouter()
-
-// Detail see: https://github.com/vuejs/vue-router/issues/1234#issuecomment-357941465
 export function resetRouter() {
-    const newRouter = createRouter()
-    router.matcher = newRouter.matcher // reset router
+    //获取所有路由
+    router.getRoutes().forEach((route) => {
+        const {name} = route;   //获取路由name
+        if (name && !whiteList.includes(name)) {      //路由不属于白名单,则删除
+            router.hasRoute(name) && router.removeRoute(name);
+        }
+    });
 }
+
+/** 路由白名单 */
+const whiteList = ['/login', '/register']
+
+const router = createRouter({
+    history: createWebHashHistory(),
+    scrollBehavior: () => ({y: 0}),
+    routes: constantRoutes,
+});
+
 
 export default router
