@@ -1,306 +1,177 @@
+<script setup>
+import {ref, onMounted, onBeforeUnmount} from 'vue'
+import {useRoute, useRouter} from "vue-router";
+import {bg, singUp, logo} from './utils/static'
+import {loginRules} from "@/utils/rules";
+import {reqLogin} from "@/api/user";
+import {storageLocal} from "@pureadmin/utils";
+import {useUserStoreHook} from "@/stores";
+import {ElMessage} from "element-plus";
+
+const router = useRouter()
+
+const ruleFormRef = ref()
+const ruleForm = ref({
+  user_name: "",
+  password: "",
+})
+const loading = ref(false)
+const isRememberMe = ref(false)
+const redirect = ref(undefined)
+
+const onLogin = async () => {
+  loading.value = true
+  await ruleFormRef.value.validate(async (valid) => {
+    if (valid) {
+      console.log(true)
+      if (isRememberMe.value) {
+        storageLocal().setItem("loginInfo", ruleForm.value)
+      } else {
+        storageLocal().removeItem("loginInfo")
+      }
+      const res = await reqLogin(ruleForm.value)
+      // console.log(ruleForm.value)
+      if (res && res.code == 0) {
+        useUserStoreHook().loginByUsername(ruleForm.value).then(res => {
+          if (res.code == 0) {
+            router.replace({path: redirect.value || '/'})
+            ElMessage({
+              type: 'success',
+              message: '登录成功'
+            })
+          }
+        })
+      } else {
+        loading.value = false
+      }
+    } else {
+      loading.value = false
+      console.log(false);
+    }
+  })
+}
+
+/** 使用公共函数，避免`removeEventListener`失效 */
+function onkeypress({code}) {
+  if (code === "Enter") {
+    onLogin(ruleFormRef.value);
+  }
+}
+
+
+onMounted(() => {
+  window.document.addEventListener("keypress", onkeypress);
+  const route = useRoute();
+  if (route.query.user_name) {
+    ruleForm.value.user_name = route.query.user_name + "";
+    isRememberMe.value = true;
+  } else {
+    if (storageLocal().getItem("loginInfo")) {
+      isRememberMe.value = true;
+      Object.assign(ruleForm, storageLocal().getItem("loginInfo"));
+    }
+  }
+});
+
+onBeforeUnmount(() => {
+  window.document.removeEventListener("keypress", onkeypress);
+});
+
+const goRegister = () => {
+  router.push("/register");
+};
+
+</script>
 
 <template>
   <div class="mi-login">
-    <img class="wave" src="@/assets/images/bbg.png">
-    <div class="login-container">
+    <img :src="bg" class="wave">
+    <div class="mi-login__container">
       <div class="img">
-        <img src="@/assets/images/signup.png">
+        <img :src="singUp">
       </div>
       <div class="login-box">
         <div class="login-form">
-          <div class="title-container">
-            <img class="logo" src="@/assets/images/logo.jpg">
-            <h3 class="outline-none">米娜小屋后台</h3>
+          <div class="title">
+            <img :src="logo" class="logo">
+            <h2 class="outline-none">米娜的小屋后台</h2>
           </div>
-          <el-form ref="loginForm" :model="loginForm" :rules="loginRules" auto-complete="on" class="login-form"
-                   label-position="left">
-
-            <el-form-item prop="username">
-        <span class="svg-container">
-          <svg-icon icon-class="user"/>
-        </span>
-              <el-input
-                  ref="username"
-                  v-model="loginForm.username"
-                  auto-complete="on"
-                  name="username"
-                  placeholder="Username"
-                  tabindex="1"
-                  type="text"
-              />
+          <el-form ref="ruleFormRef" :model="ruleForm" :rules="loginRules">
+            <el-form-item prop="user_name">
+              <el-input v-model="ruleForm.user_name" clearable placeholder="账号">
+                <template #prefix>
+                  <el-icon>
+                    <UserFilled/>
+                  </el-icon>
+                </template>
+              </el-input>
             </el-form-item>
-
             <el-form-item prop="password">
-        <span class="svg-container">
-          <svg-icon icon-class="password"/>
-        </span>
-              <el-input
-                  :key="passwordType"
-                  ref="password"
-                  v-model="loginForm.password"
-                  :type="passwordType"
-                  auto-complete="on"
-                  name="password"
-                  placeholder="Password"
-                  tabindex="2"
-                  @keyup.enter.native="handleLogin"
-              />
-              <span class="show-pwd" @click="showPwd">
-          <svg-icon :icon-class="passwordType === 'password' ? 'eye' : 'eye-open'"/>
-        </span>
+              <el-input v-model="ruleForm.password" clearable placeholder="密码" show-password>
+                <template #prefix>
+                  <el-icon>
+                    <Lock/>
+                  </el-icon>
+                </template>
+              </el-input>
             </el-form-item>
-
-            <el-button :loading="loading" style="width:100%;margin-bottom:30px;" type="primary"
-                       @click.native.prevent="handleLogin">Login
-            </el-button>
-
-            <div class="tips">
-              <span style="margin-right:20px;">username: admin</span>
-              <span> password: any</span>
+            <div class="register">
+              <el-checkbox v-model="isRememberMe">记住我</el-checkbox>
+              <div>没有账号？<span class="line" @click="goRegister">去注册</span></div>
             </div>
-
+            <el-button class="loginBtn" size="default" type="primary" @click="onLogin()">立即登录</el-button>
           </el-form>
         </div>
-
       </div>
     </div>
   </div>
 </template>
 
-<script>
-import {validUsername} from '@/utils/validate'
-
-export default {
-  name: 'Login',
-  data() {
-    const validateUsername = (rule, value, callback) => {
-      if (!validUsername(value)) {
-        callback(new Error('Please enter the correct user name'))
-      } else {
-        callback()
-      }
-    }
-    const validatePassword = (rule, value, callback) => {
-      if (value.length < 6) {
-        callback(new Error('The password can not be less than 6 digits'))
-      } else {
-        callback()
-      }
-    }
-    return {
-      loginForm: {
-        username: 'admin',
-        password: '111111'
-      },
-      loginRules: {
-        username: [{required: true, trigger: 'blur', validator: validateUsername}],
-        password: [{required: true, trigger: 'blur', validator: validatePassword}]
-      },
-      loading: false,
-      passwordType: 'password',
-      redirect: undefined
-    }
-  },
-  watch: {
-    $route: {
-      handler: function (route) {
-        this.redirect = route.query && route.query.redirect
-      },
-      immediate: true
-    }
-  },
-
-  methods: {
-    showPwd() {
-      if (this.passwordType === 'password') {
-        this.passwordType = ''
-      } else {
-        this.passwordType = 'password'
-      }
-      this.$nextTick(() => {
-        this.$refs.password.focus()
-      })
-    },
-    handleLogin() {
-      this.$refs.loginForm.validate(valid => {
-        if (valid) {
-          this.loading = true
-          this.$store.dispatch('user/login', this.loginForm).then(() => {
-            this.$router.push({path: this.redirect || '/'})
-            this.loading = false
-          }).catch(() => {
-            this.loading = false
-          })
-        } else {
-          console.log('error submit!!')
-          return false
-        }
-      })
-    }
-  }
-}
-</script>
-
-<style lang="scss">
-/* 修复input 背景不协调 和光标变色 */
-/* Detail see https://github.com/PanJiaChen/vue-element-admin/pull/927 */
-
-$bg: #283443;
-$light_gray: #fff;
-$cursor: #fff;
-
-@supports (-webkit-mask: none) and (not (cater-color: $cursor)) {
-  .login-container .el-input input {
-    color: $cursor;
-  }
-}
-
-/* reset element-ui css */
-.login-container {
-  .el-input {
-    display: inline-block;
-    height: 47px;
-    width: 85%;
-
-    input {
-      background: transparent;
-      border: 0px;
-      -webkit-appearance: none;
-      border-radius: 0px;
-      padding: 12px 5px 12px 15px;
-      color: $light_gray;
-      height: 47px;
-      caret-color: $cursor;
-
-      &:-webkit-autofill {
-        box-shadow: 0 0 0px 1000px $bg inset !important;
-        -webkit-text-fill-color: $cursor !important;
-      }
-    }
-  }
-
-  .el-form-item {
-    border: 1px solid rgba(255, 255, 255, 0.1);
-    background: rgba(0, 0, 0, 0.1);
-    border-radius: 5px;
-    color: #454545;
-  }
-}
+<style scoped>
+@import url("@/styles/login.css");
 </style>
 
 <style lang="scss" scoped>
-$bg: #2d3a4b;
-$dark_gray: #889aa4;
-$light_gray: #eee;
-$light_black: #7c7878;
+:deep(.el-input-group__append, .el-input-group__prepend) {
+  padding: 0;
+}
 
-.wave {
-  position: fixed;
-  height: 100%;
+:deep(.el-input__wrapper) {
+  border-radius: 20px;
+  height: 2.5rem;
+}
+
+.title {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 360px;
+  height: 50px;
+  margin-bottom: 20px;
+}
+
+.register {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  font-size: 0.8em;
+  color: #676767;
+  padding: 0 10px;
+
+  :deep(.el-checkbox.el-checkbox--large .el-checkbox__label) {
+    font-size: 0.8em;
+    color: #676767;
+  }
+}
+
+.line {
+  cursor: pointer;
+  text-decoration: underline;
+}
+
+.loginBtn {
   width: 100%;
-  left: 0;
-  button: 0;
-  z-index: -1;
-}
-
-.logo {
-  width: 50px;
-  margin-right: 10px;
-}
-
-.img {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-}
-
-.img img {
-  width: 500px;
-}
-
-.login-box {
-  display: flex;
-  align-items: center;
-  text-align: center;
-}
-
-.login-container {
-  //background-color: $bg;
-  overflow: hidden;
-  width: 100vw;
-  height: 100vh;
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  grid-gap: 10rem;
-  padding: 0 2rem;
-
-  .login-form {
-    position: relative;
-    width: 520px;
-    max-width: 100%;
-    margin: 0 auto;
-    overflow: hidden;
-  }
-
-  .tips {
-    font-size: 14px;
-    color: #fff;
-    margin-bottom: 10px;
-
-    span {
-      &:first-of-type {
-        margin-right: 16px;
-      }
-    }
-  }
-
-  .svg-container {
-    padding: 6px 5px 6px 15px;
-    color: $dark_gray;
-    vertical-align: middle;
-    width: 30px;
-    display: inline-block;
-  }
-
-  .title-container {
-    position: relative;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    text-align: center;
-    width: 100%;
-    height: 50px;
-    font-size: 26px;
-    color: $light_black;
-    margin-bottom: 20px;
-
-    //.title {
-    //  font-size: 26px;
-    //  color: $light_gray;
-    //  margin: 0px auto 40px auto;
-    //  text-align: center;
-    //  font-weight: bold;
-    //}
-  }
-
-  .show-pwd {
-    position: absolute;
-    right: 10px;
-    top: 7px;
-    font-size: 16px;
-    color: $dark_gray;
-    cursor: pointer;
-    user-select: none;
-  }
-}
-
-@media screen and (max-width: 1180px) {
-  .login-container {
-    grid-gap: 9rem;
-  }
-
-  .logo {
-    width: 50px;
-  }
+  border-radius: 20px;
+  margin-top: 20px;
 }
 </style>
