@@ -1,6 +1,6 @@
-import bg from '@/assets/images/bbg.png'
-import logo from '@/assets/images/logo.jpg'
-import singIn from '@/assets/svg/signin.svg?component'
-import into from '@/assets/svg/signup.svg?component'
+import bg from "@/assets/images/bbg.png";
+import logo from "@/assets/images/logo.jpg"
+import singIn from "@/assets/svg/signin.svg?component";
+import singUp from "@/assets/svg/signup.svg?component"
 
-export {bg, singIn, into, logo}
+export {bg, singIn, singUp, logo}
