@@ -43,7 +43,7 @@ module.exports = {
                 target: 'http://localhost:8888',
                 changeOrigin: true,
                 pathRewrite: {
-                    '^/api': '/'
+                    '^/api': ''
                 }
             }
         }
